+++ nautilus_trader/adapters/interactive_brokers/client/client.py
@@ -95,7 +95,7 @@ class InteractiveBrokersClient(
         port: int = 7497,
         client_id: int = 1,
         fetch_all_open_orders: bool = False,
-        request_timeout: int = 60,
+        request_timeout_secs: int = 60,
     ) -> None:
         super().__init__(
             clock=clock,
@@ -111,7 +111,7 @@ class InteractiveBrokersClient(
         self._port = port
         self._client_id = client_id
         self._fetch_all_open_orders = fetch_all_open_orders
-        self._request_timeout = request_timeout
+        self._request_timeout_secs = request_timeout_secs
 
         # TWS API
         self._eclient: EClient = EClient(