+++ nautilus_trader/adapters/interactive_brokers/client/contract.py
@@ -70,13 +70,13 @@ class InteractiveBrokersClientContractMixin(BaseMixin):
 
             return await self._await_request(
                 request,
-                self._request_timeout,
+                self._request_timeout_secs,
                 suppress_timeout_warning=True,
             )
         else:
             return await self._await_request(
                 request,
-                self._request_timeout,
+                self._request_timeout_secs,
                 suppress_timeout_warning=True,
             )
 
@@ -113,7 +113,7 @@ class InteractiveBrokersClientContractMixin(BaseMixin):
 
             request.handle()
 
-            return await self._await_request(request, self._request_timeout)
+            return await self._await_request(request, self._request_timeout_secs)
         else:
             self._log.info(f"Request already exist for {request}")
             return None
@@ -154,7 +154,7 @@ class InteractiveBrokersClientContractMixin(BaseMixin):
 
             request.handle()
 
-            return await self._await_request(request, self._request_timeout)
+            return await self._await_request(request, self._request_timeout_secs)
         else:
             self._log.info(f"Request already exist for {request}")
             return None