+++ nautilus_trader/adapters/interactive_brokers/factories.py
@@ -54,7 +54,7 @@ def get_cached_ib_client(
     client_id: int = 1,
     dockerized_gateway: DockerizedIBGatewayConfig | None = None,
     fetch_all_open_orders: bool = False,
-    request_timeout: int = 60,
+    request_timeout_secs: int = 60,
 ) -> InteractiveBrokersClient:
     """
     Retrieve or create a cached InteractiveBrokersClient using the provided key.
@@ -88,7 +88,7 @@ def get_cached_ib_client(
     fetch_all_open_orders : bool, default False
         If True, uses reqAllOpenOrders to fetch orders from all API clients and TWS GUI.
         If False, uses reqOpenOrders to fetch only orders from current client ID session.
-    request_timeout : int, default 60
+    request_timeout_secs : int, default 60
         The timeout (seconds) to wait for request responses (contract details, etc.).
 
     Returns
@@ -129,7 +129,7 @@ def get_cached_ib_client(
             port=port,
             client_id=client_id,
             fetch_all_open_orders=fetch_all_open_orders,
-            request_timeout=request_timeout,
+            request_timeout_secs=request_timeout_secs,
         )
         client.start()
         IB_CLIENTS[client_key] = client
@@ -228,7 +228,7 @@ class InteractiveBrokersLiveDataClientFactory(LiveDataClientFactory):
             port=config.ibg_port,
             client_id=config.ibg_client_id,
             dockerized_gateway=config.dockerized_gateway,
-            request_timeout=config.request_timeout,
+            request_timeout_secs=config.request_timeout_secs,
         )
 
         # Get instrument provider singleton
@@ -302,7 +302,7 @@ class InteractiveBrokersLiveExecClientFactory(LiveExecClientFactory):
             client_id=config.ibg_client_id,
             dockerized_gateway=config.dockerized_gateway,
             fetch_all_open_orders=config.fetch_all_open_orders,
-            request_timeout=config.request_timeout,
+            request_timeout_secs=config.request_timeout_secs,
         )
 
         # Get instrument provider singleton